Use rayon instead of custom thread pool
