+++ datafusion-scheduler/Cargo.toml
@@ -38,12 +38,11 @@ path = "src/lib.rs"
 ahash = { version = "0.7", default-features = false }
 arrow = { version = "11" }
 async-trait = "0.1"
-crossbeam-channel = "0.5"
-crossbeam-deque = "0.8"
 datafusion = { path = "../datafusion", version = "7.0.0" }
 futures = "0.3"
 log = "0.4"
 parking_lot = "0.12"
+rayon = "1.5"
 
 [dev-dependencies]
 criterion = "0.3"
